Add notes and examples for justify-content property.
Worked on how the justify-content property allows you to align flex-items along the main axis. Learned that the justify-content property will be different depending on whether the main axis is horizontal (flex-direction: row) or vertical (flex-direction: column). If the main axis runs vertically, you need to ensure the flex-container has a height.