Add custom image and imagemin dependency to build.
The sample file had an image being pulled from a website, and it took to long to load on refreshes so I added my own stock image. I figured I would add a build step in the gulpfile for this as well, so hooked up imagemin (although it doesn't seem to do anything...).