Add notes and refactor html/css to lower specificity.
I realized I didn't like the way Wes had some of the classes setup in the html, and also I didn't like the way some of the style rules of my own were too specific. Specifity can be a bad thing, so it's good to try and mitigate it by choosing the most general style rule you can. I refactored my html and css to have lower specificity, and to make a little more sense to me.