Refactor JavaScript to have toggle functionality for nav button.
While the script I wrote would show the nav menu when the button was clicked, it would not hide the menu when I clicked the button again. This is because I was merely adding a class to the classList of the nav bar to get it to display everytime the button was clicked. If you want a toggle to happen, you need to add the .open class when the button is clicked if it is NOT present, and remove the class if it IS present. the classList property has a handy toggle() function that does just that!